add meaningful unit test
