+++ app/src/main/java/ch/epfl/favo/exceptions/NoPositionFoundException.java
@@ -1,0 +1,7 @@
+package ch.epfl.favo.exceptions;
+
+public class NoPositionFoundException extends RuntimeException {
+    public NoPositionFoundException(String s) {
+        super(s);
+    }
+}
