+++ app/src/main/java/ch/epfl/favo/presenter/MainActivity.java
@@ -29,7 +29,6 @@ public class MainActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
-
         //Use tabs.
         setUpViewPager();
