+++ app/src/main/res/layout/fragment_tab1.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presenter.tabs.Tab1">
+    tools:context=".presenter.tabs.MapsTab">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
